% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_set-class-methods.R
\name{as.data.frame.Response_set}
\alias{as.data.frame.Response_set}
\title{Convert a \code{\link{Response_set-class}} object into a long
format \code{data.frame}}
\usage{
\method{as.data.frame}{Response_set}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{Response_set-class}} object}

\item{row.names}{\code{NULL} or a character vector giving the row names for
the data frame. Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and converting
column names}

\item{...}{additional arguments}
}
\description{
Convert a \code{\link{Response_set-class}} object into a long
format \code{data.frame}
}
\author{
Emre Gonulates
}
