% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class-methods.R
\name{as.Itempool}
\alias{as.Itempool}
\title{Coerce a given object to \code{\link{Itempool-class}} object}
\usage{
as.Itempool(...)
}
\arguments{
\item{...}{The object that is desired to be converted to an  'Itempool'
object. Also additional arguments related to the Itempool.}
}
\value{
An \code{\link{Itempool-class}} object.
}
\description{
This function is a wrapper for \code{\link{itempool}} function.
It is recommended to use that function.
}
\seealso{
\code{\link{itempool}}
}
\author{
Emre Gonulates
}
