% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class-methods.R
\name{as.list.Itempool}
\alias{as.list.Itempool}
\title{This function converts Itempool objects to a list object}
\usage{
\method{as.list}{Itempool}(x, ...)
}
\arguments{
\item{x}{an \code{\link{Itempool-class}} to be coerced to a list object}

\item{...}{Additional parameters to be passed to the function.}
}
\value{
A list object with elements from 'Item' class.
}
\description{
This function converts Itempool objects to a list object
}
\examples{
item1 <- item(a = 1.12, b = -2.1, c = 0.28)
item2 <- item(a = 2, b = 3.2, c = 0.21)

ip1 <- c(item1, item2)
as.list(ip1)

}
\author{
Emre Gonulates
}
