% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_set-class-methods.R
\name{as.list.Response_set}
\alias{as.list.Response_set}
\title{This function converts Response_set objects to a list object}
\usage{
\method{as.list}{Response_set}(x, ...)
}
\arguments{
\item{x}{an \code{\link{Response_set-class}} to be coerced to a list object}

\item{...}{Additional parameters to be passed to the function.}
}
\value{
A list object with elements from \code{\link{Response-class}}
  objects.
}
\description{
This function converts Response_set objects to a list object
}
\author{
Emre Gonulates
}
