% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_sim_helper_functions.R
\name{c.cat_design}
\alias{c.cat_design}
\title{Concatenate 'cat_design' objects}
\usage{
\method{c}{cat_design}(x, ...)
}
\arguments{
\item{x}{A \code{cat_design} class object.}

\item{...}{Remaining \code{cat_design} class objects.}
}
\value{
A list of \code{cat_design} objects.
}
\description{
Concatenate 'cat_design' objects
}
\examples{

ip <- generate_ip(n = 20)
cd1 <- create_cat_design(ip = ip,
                         termination_rule = c('max_item'),
                         termination_par = list(max_item = 5))
cd2 <- create_cat_design(ip = ip,
                         termination_rule = c('max_item'),
                         termination_par = list(max_item = 9))
cd <- c(cd1, cd2)
}
\author{
Emre Gonulates
}
