% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_set-class-methods.R
\name{$,Response_set-method}
\alias{$,Response_set-method}
\title{Get slots of the a \code{\link{Response_set-class}} object.}
\usage{
\S4method{$}{Response_set}(x, name)
}
\arguments{
\item{x}{An \code{\link{Response_set-class}} object.}

\item{name}{Name of the parameter.
Available values:
\describe{
  \item{\strong{\code{'response_list'}}}{Extract Response objects
    as a list.}
  \item{\strong{\code{'item_id'}}}{Extract unique list of item IDs that
    are in the response set.}
  \item{\strong{\code{'testlet_id'}}}{Extract unique list of testlet IDs
    that are in the response set.}
  \item{\strong{\code{'misc'}}}{Extract 'misc' field.}
  \item{\strong{\code{'score'}}}{Return a score matrix of responses}
  \item{\strong{\code{'raw_response'}}}{Return a raw score matrix of
    responses}
}}
}
\value{
See the 'name' argument above for possible return values.
}
\description{
Get slots of the a \code{\link{Response_set-class}} object.
}
\examples{
resp <- sim_resp(ip = generate_ip(), theta = rnorm(5),
                 output = "response_set")
resp$response_list

}
\author{
Emre Gonulates
}
