% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-class-methods.R
\name{$<-,Item-method}
\alias{$<-,Item-method}
\title{Set values to parameters or components of \code{\link{Item-class}} object}
\usage{
\S4method{$}{Item}(x, name) <- value
}
\arguments{
\item{x}{An \code{\link{Item-class}} object.}

\item{name}{Name of the parameter or component.}

\item{value}{The new value that will be assigned.}
}
\value{
This operation will not return anything.
}
\description{
Set values to parameters or components of \code{\link{Item-class}} object
}
\examples{
item <- new("Item", model =  "3PL", id = 'item23', content = 'Geometry',
            misc = list(enemies = c("item1", "item2")),
            parameters = list(b = 2, c = .12, a = 1.2, D = 1))
item$a <- 2
item$D <- 1.7
item$id <- "Itm-111"
item$content <- 'Algebra'
item$misc <- list(enemies = c("item5"))
}
\author{
Emre Gonulates
}
