% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_set-class-methods.R
\name{$<-,Response_set-method}
\alias{$<-,Response_set-method}
\title{Set values to components of 'Response_set' class objects}
\usage{
\S4method{$}{Response_set}(x, name) <- value
}
\arguments{
\item{x}{\code{\link{Response_set-class}} object.}

\item{name}{Name of the parameter or component. Currently only
\code{response_list}, \code{misc} are available.}

\item{value}{The new value that will be assigned.}
}
\value{
This operation will return an \code{\link{Response_set-class}}
  object.
}
\description{
Set values to components of 'Response_set' class objects
}
\author{
Emre Gonulates
}
