% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itempool-class-methods.R
\name{$<-,Testlet-method}
\alias{$<-,Testlet-method}
\title{Set values to parameters or components of \code{\link{Testlet-class}} object}
\usage{
\S4method{$}{Testlet}(x, name) <- value
}
\arguments{
\item{x}{An \code{\link{Testlet-class}} object.}

\item{name}{Name of the parameter or component.}

\item{value}{The new value that will be assigned.}
}
\value{
This operation will not return anything.
}
\description{
Set values to parameters or components of \code{\link{Testlet-class}} object
}
\examples{
tlt <- generate_testlet()
tlt$testlet_id <- "New-Testlet-ID-111"
tlt$content <- "Algebra"
# Set all misc fields like this
tlt$misc <- list(passage_text = "This is a reading passage.",
                 passage_lexile = 450)

tlt$passage_text
# Add a misc field
tlt$passage_language <- "En-US"

# Remove a misc field
tlt$passage_language <- NULL

}
\author{
Emre Gonulates
}
