% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class-methods.R
\name{convert_model}
\alias{convert_model}
\alias{convert_model,Item-method}
\alias{convert_model,Itempool-method}
\alias{convert_model,Testlet-method}
\title{Convert model parameters from one model to another}
\usage{
convert_model(ip, target_model = "3PL")

\S4method{convert_model}{Item}(ip, target_model = "3PL")

\S4method{convert_model}{Itempool}(ip, target_model = "3PL")

\S4method{convert_model}{Testlet}(ip, target_model = "3PL")
}
\arguments{
\item{ip}{An \code{\link{Item-class}} or \code{\link{Itempool-class}} object}

\item{target_model}{The target model that the conversion will be made.}
}
\value{
An 'Item' object with new model parameters will be returned.
}
\description{
This is especially handy for converting IRT models with less parameters
(such as 1 parameter logistic model) to higher dimensional models such as
three parameter logistic model.
}
\author{
Emre Gonulates

Emre Gonulates

Emre Gonulates

Emre Gonulates
}
\keyword{internal}
