% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{dif}
\alias{dif}
\title{Evaluate Differential Item Functioning (DIF) of a test}
\usage{
dif(resp, group, focal_name, ip = NULL, type = "mh")
}
\arguments{
\item{resp}{A vector of item responses.}

\item{group}{Group membership}

\item{focal_name}{In the group variable, the value that represents the focal
group.}

\item{ip}{An \code{\link{Itempool-class}} object.}

\item{type}{The type of the DIF method.}
}
\value{
A data.frame of DIF values.
}
\description{
\code{dif} evaluates Differential Item Functioning (DIF) of a test.
}
\author{
Emre Gonulates
}
