% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testlet-class-methods.R
\name{.print.Testlet}
\alias{.print.Testlet}
\title{Print a \code{\link{Testlet-class}} object}
\usage{
.print.Testlet(x, ...)
}
\arguments{
\item{x}{An \code{\link{Testlet-class}} object to be printed.}

\item{...}{Passed parameters.}
}
\description{
Print a \code{\link{Testlet-class}} object
}
\author{
Emre Gonulates
}
\keyword{internal}
