\name{elections}
\alias{elections}
\title{Elections Data}
\description{
Data 'elections' contains 2502 individuals and ...
}
\usage{data(elections)}  
\format{A data frame containing 2502 observations on ... variables:
\tabular{rl}{
        [,1] \tab "govguareatsleep - description ..."\cr
        [,2] \tab "govtakecarewhocant - The government should take care of those who can't take of themselves"\cr
        [,3] \tab "govhelpneedy - description ..."\cr
        [,4] \tab "improveblackpos - description ..."\cr
        [,5] \tab "equalOPP - description ..."\cr
        [,6] \tab "poortoodep - description ..."\cr
        [,7] \tab "richgetricher - description ..."\cr
        [,8] \tab "govwasteful - description ..."\cr
        [,9] \tab "unborn - description ..."\cr
        [,10] \tab "womentradrole - description ..."\cr
        [,11] \tab "peacethrustrength - description ..."\cr
        [,12] \tab "searchterrorists - description ..."\cr
        [,13] \tab "bandangerousbooks - description ..."\cr
        [,14] \tab "firegayteachers - description ..."\cr
        [,15] \tab "xratedOK - description ..."\cr
        [,16] \tab "age - description ..."\cr
        [,17] \tab "agesq.01 - description ..."\cr
        [,18] \tab "black - description ..."\cr
        [,19] \tab "bornagain - description ..."\cr
        [,20] \tab "blackbornagain - description ..."\cr
        [,21] \tab "rel.catholic - description ..."\cr
        [,22] \tab "rel.nonchr - description ..."\cr
        [,23] \tab "ed.cat1 - description ..."\cr
        [,24] \tab "ed.cat3 - description ..."\cr
        [,25] \tab "ed.cat4 - description ..."\cr
        [,26] \tab "ed.cat5 - description ..."\cr
        [,27] \tab "income.1 - description ..."\cr
        [,28] \tab "income.3 - description ..."\cr
        [,29] \tab "income.4 - description ..."\cr
        [,30] \tab "income.dk - description ..."\cr
        [,31] \tab "parent - description ..."\cr
        [,32] \tab "hispanic - spanish origins ..."\cr
        [,33] \tab "female - description ..."
        }   
}           
\references{}
\keyword{datasets}   