\name{etF}
\alias{etF}
\title{etF}
\description{ etF computes the distribution function of an exponentially tilted uniform or beta density by numerical integration }
\arguments{   
   \item{u}{to set and describe}
   \item{tvec}{ to set and describe}
   \item{scheme}{default NULL}
   \item{nq}{8}
   \item{beta}{beta}
}
\value{ Function returns a... }
\author{Richard Spady}
\examples{}   
\keyword{} 