% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{GPCM2-class}
\alias{GPCM2-class}
\title{Reparametrized Generalized Partial Credit Model}
\description{
Reparametrized Generalized Partial Credit Model
}
\section{Slots}{

\describe{
\item{\code{a}}{Item discrimination parameter}

\item{\code{b}}{Overall location parameter}

\item{\code{d}}{A vector of threshold parameters}

\item{\code{D}}{Scaling constant}

\item{\code{se_a}}{Standard error of item discrimination parameter}

\item{\code{se_b}}{Standard error of overall location parameter}

\item{\code{se_d}}{A vector of standard error of item threshold parameters}
}}

\author{
Emre Gonulates
}
