% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{GRM-class}
\alias{GRM-class}
\title{Graded Response Model}
\description{
Graded Response Model
}
\section{Slots}{

\describe{
\item{\code{a}}{Item discrimination parameter}

\item{\code{b}}{A vector of threshold parameters}

\item{\code{D}}{Scaling constant}

\item{\code{se_a}}{Standard error of item discrimination parameter}

\item{\code{se_b}}{A vector of standard error of item threshold parameters}
}}

\author{
Emre Gonulates
}
