% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-irt.R
\docType{package}
\name{irt-package}
\alias{irt}
\alias{irt-package}
\title{irt: Item Response Theory and Computerized Adaptive Testing Functions}
\description{
A collection of Item Response Theory (IRT) and Computerized Adaptive Testing (CAT) functions that are used in psychometrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/egonulates/irt}
  \item Report bugs at \url{https://github.com/egonulates/irt/issues}
}

}
\author{
\strong{Maintainer}: Emre Gonulates \email{egonulates@gmail.com} (\href{https://orcid.org/0000-0002-3834-3266}{ORCID})

}
\keyword{internal}
