% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{1PL-class}
\alias{1PL-class}
\title{One-Parameter Logistic IRT model}
\description{
One-Parameter Logistic IRT model
}
\section{Slots}{

\describe{
\item{\code{b}}{Item difficulty parameter}

\item{\code{D}}{Scaling constant}

\item{\code{se_b}}{Standard error of item difficulty parameter}
}}

\author{
Emre Gonulates
}
