% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{2PL-class}
\alias{2PL-class}
\title{Two-Parameter Logistic IRT model}
\description{
Two-Parameter Logistic IRT model
}
\section{Slots}{

\describe{
\item{\code{a}}{Item discrimination parameter}

\item{\code{b}}{Item difficulty parameter}

\item{\code{D}}{Scaling constant}

\item{\code{se_a}}{Standard error of item discrimination parameter}

\item{\code{se_b}}{Standard error of item difficulty parameter}
}}

\author{
Emre Gonulates
}
