% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{3PL-class}
\alias{3PL-class}
\title{Three-Parameter Logistic IRT model}
\description{
Three-Parameter Logistic IRT model
}
\section{Slots}{

\describe{
\item{\code{a}}{Item discrimination parameter}

\item{\code{b}}{Item difficulty parameter}

\item{\code{c}}{Guessing parameter}

\item{\code{D}}{Scaling constant}

\item{\code{se_a}}{Standard error of item discrimination parameter}

\item{\code{se_b}}{Standard error of item difficulty parameter}

\item{\code{se_c}}{Standard error of guessing  parameter}
}}

\author{
Emre Gonulates
}
