% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-class-methods.R,
%   R/Itempool-class-methods.R, R/Testlet-class-methods.R
\name{is.Item}
\alias{is.Item}
\alias{is.Itempool}
\alias{is.Testlet}
\title{Check whether an object is an \code{\link{Item-class}}}
\usage{
is.Item(x)

is.Itempool(x)

is.Testlet(x)
}
\arguments{
\item{x}{an object that is checked for being a member of 'Testlet' class}
}
\description{
Check whether an object is an \code{\link{Item-class}}

Check whether an object is an  \code{\link{Itempool-class}} object

Check whether an object is a \code{\link{Testlet-class}} object
}
\examples{
i1 <- item(a = 1, b = 2)
is.Item(i1)
# Alternatively:
is(i1, "Item")

# Not an item:
is.Item("abc")

}
\author{
Emre Gonulates

Emre Gonulates

Emre Gonulates
}
