% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-class-methods.R
\name{item}
\alias{item}
\title{Create an \code{Item} object}
\usage{
item(
  ...,
  model = NULL,
  id = NULL,
  parameters = NULL,
  se_parameters = NULL,
  content = NULL,
  misc = NULL
)
}
\arguments{
\item{...}{The item parameter arguments.}

\item{model}{The model that item \code{parameters} represents. Currently
model can be:
1PL, 2PL, 3PL, 4PL, M1PL, M2PL and M3PL, GRM, PCM or GPCM.
Ideally, a model should be specified for the construction of an
\code{\link{Item-class}} object.}

\item{id}{Item id. Default value is \code{NULL}.}

\item{parameters}{A list containing numeric vectors that represent item
parameters. Depending on the model these can change.}

\item{se_parameters}{Standard error of item parameters.}

\item{content}{Content information for item.}

\item{misc}{This slot is a list where one can put any information about
the item. For example, one can enter the id's of the enemies of the current
item as \code{misc = list(enemies = c("i1", i2))}. Or, one can enter
Sympson-Hetter exposure control parameter K:
\code{misc = list(sympson_hetter_k = .75)}.}
}
\value{
An \code{Item} class object.
}
\description{
This function is used for creating \code{\link{Item-class}}
  objects.
}
\examples{
# Create 2PL item:
item(a = 1.2, b = -0.94)
item(a = 1.2, b = -0.94, model = "2PL")
# Specify scaling constant D:
item(a = 1.2, b = -0.94, D = 1.7)

# Add additional item specifications:
# Add id
item(a = 1.2, b = -0.94, id = "My-Item-1")
# Add content
item(a = 1.2, b = -0.94, id = "My-Item-1", content = "Geometry")
# Add additional parameter
item(a = 1.2, b = -0.94, misc = list(sympson_hetter_k = 1))
# Add any argument to 'misc' field
i1 <- item(a = 1.2, b = -0.94, id = "item1", content = "Earth Science",
           misc = list(key = "C", operational = TRUE, type = "MC",
                       enemies = c("i2", "i3")))
# Access fields
i1$misc
i1$misc$key
i1$misc$operational
i1$misc$enemies
i1$a
i1$b
i1$D
i1$parameters
i1$id
i1$content


# Rasch Model
item(b = 1.2)
item(b = 1.2, model = "Rasch")

# 1PL model:
item(b = 1.2, model = "1PL")
item(b = 1.2, D = 1)

# 3PL model:
item(a = 0.92, b = 2.7, c = 0.17)
item(a = 0.92, b = 2.7, c = 0.17, model = "3PL")
item(a = 0.92, b = 2.7, c = 0.17, D = 1.7, model = "3PL")

# 4PL model:
item(a = 0.92, b = 2.7, c = 0.17, d = 0.98)
item(a = 0.92, b = 2.7, c = 0.17, d = 0.98, model = "4PL")
item(a = 0.92, b = 2.7, c = 0.17, d = 0.92, D = 1.7, model = "4PL")
item(parameters = list(a = 0.92, b = 2.7, c = 0.17, d = 0.92, D = 1.7),
     model = "4PL")

# Create a GRM model
item(a = 1.9, b = c(-1, 0.82, 1.5), model = "GRM")
item(parameters = list(a = 1.9, b = c(-1, 2), D = 1), model = "GRM")

# Create a GPCM model
item(a = 1.9, b = c(-1.6, -0.09, 1.25), model = "GPCM")
item(parameters = list(a = 1.9, b = c(-1, 2), D = 1), model = "GPCM")

# Create a GPCM2 model (Reparametrized GPCM model)
item(a = 1.9, b = 0.65, d = c(-1.6, -0.09, 1.25), model = "GPCM2")
item(parameters = list(a = 1.9, b = 0.65, d = c(-1.6, -0.09, 1.25), D = 1.7),
     model = "GPCM2")

# Create a PCM model
item(b = c(-0.7, 0.72, 1.9), model = "PCM")
item(parameters = list(b = c(-1, 2)), model = "PCM")

# Add additional arguments to items
i1 <- item(a = 1.2, b = 2)
i1 <- item(i1, id = "new_item_id", content = "Algebra")
}
\author{
Emre Gonulates
}
