% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class.R
\docType{class}
\name{Itempool-class}
\alias{Itempool-class}
\title{An S4 class to represent an Itempool}
\description{
\code{\link{Itempool-class}} is a class to represent an item pool. This class is composed
of the collection of 'Item' class objects.
}
\section{Slots}{

\describe{
\item{\code{item_list}}{The list of items that are 'Item' class}

\item{\code{misc}}{A list of additional parameters for the item pool. For example,
one can put the calibration date of the item pool as
\code{misc = list(calibration_date = as.Date("2020-01-17"))}.}
}}

\author{
Emre Gonulates
}
