% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{M2PL-class}
\alias{M2PL-class}
\title{Multidimensional Two-Parameter Logistic Model}
\description{
Multidimensional Two-Parameter Logistic Model
}
\section{Slots}{

\describe{
\item{\code{a}}{Slope Parameters}

\item{\code{d}}{Intercept Parameter}

\item{\code{D}}{Scaling constant}

\item{\code{se_a}}{Standard errors of slope parameters}

\item{\code{se_d}}{Standard error of intercept parameter}
}}

\author{
Emre Gonulates
}
