% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{M3PL-class}
\alias{M3PL-class}
\title{Multidimensional Three-Parameter Logistic Model}
\description{
Multidimensional Three-Parameter Logistic Model
}
\section{Slots}{

\describe{
\item{\code{a}}{Slope Parameters}

\item{\code{d}}{Intercept Parameter}

\item{\code{c}}{Pseudo-Guessing Parameter}

\item{\code{D}}{Scaling constant}

\item{\code{se_a}}{Standard errors of slope parameters}

\item{\code{se_d}}{Standard error of intercept parameter}

\item{\code{se_c}}{Standard error of pseudo-guessing  parameter}
}}

\author{
Emre Gonulates
}
