% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_score.R
\name{max_score}
\alias{max_score}
\alias{max_score,Item-method}
\alias{max_score,Itempool-method}
\title{Calculate the maximum possible score}
\usage{
max_score(ip, resp = NULL, sum = TRUE)

\S4method{max_score}{Item}(ip, resp = NULL, sum = TRUE)

\S4method{max_score}{Itempool}(ip, resp = NULL, sum = TRUE)
}
\arguments{
\item{ip}{An \code{\link{Item-class}} or an \code{\link{Itempool-class}}
object containing the item parameters.}

\item{resp}{A \code{\link{Response-class}} or
\code{\link{Response_set-class}} object.}

\item{sum}{If \code{TRUE}, when \code{ip} is an \code{\link{Itempool-class}}
object the individual maximum possible scores of items will be summed.
This argument will be ignored when \code{resp} is not \code{NULL}.}
}
\value{
Maximum possible score of each item
}
\description{
Calculate the maximum possible score
}
\author{
Emre Gonulates
}
