% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{PCM-class}
\alias{PCM-class}
\title{Partial Credit Model}
\description{
Partial Credit Model
}
\section{Slots}{

\describe{
\item{\code{b}}{A vector of threshold parameters}

\item{\code{se_b}}{A vector of standard error of item threshold parameters}
}}

\author{
Emre Gonulates
}
