% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_analysis.R
\name{point_biserial}
\alias{point_biserial}
\title{Calculate point-biserial correlation}
\usage{
point_biserial(score, total_score)
}
\arguments{
\item{score}{Item scores of each examinee for which point-biserial
correlation will be calculated}

\item{total_score}{Total score of each examinee}
}
\value{
Point-biserial correlation value
}
\description{
Calculate point-biserial correlation
}
\author{
Emre Gonulates
}
\keyword{internal}
