% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class-methods.R
\name{print,Itempool-method}
\alias{print,Itempool-method}
\title{Show an \code{\link{Itempool-class}} object}
\usage{
\S4method{print}{Itempool}(x, ...)
}
\arguments{
\item{x}{An \code{\link{Itempool-class}} object that will be showed.}

\item{...}{Additional arguments. For example, an argument \code{n = 14},
will print 14 items to the console.}
}
\description{
Show an \code{\link{Itempool-class}} object
}
\examples{
ip <- generate_ip(n = 42)
print(ip)
print(ip, n = 3)
print(ip, n = 12)
print(ip, n = Inf)

}
\author{
Emre Gonulates
}
\keyword{internal}
