% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-class-methods.R
\name{print,Response-method}
\alias{print,Response-method}
\title{Show an \code{\link{Response-class}} object}
\usage{
\S4method{print}{Response}(x, ...)
}
\arguments{
\item{x}{A \code{\link{Response-class}} object that will be printed.}

\item{...}{Additional arguments. For example, an argument \code{n = 14},
will print 14 rows of the responses/scores.}
}
\description{
Show an \code{\link{Response-class}} object
}
\author{
Emre Gonulates
}
\keyword{internal}
