% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_set-class-methods.R
\name{print,Response_set-method}
\alias{print,Response_set-method}
\title{Print a \code{\link{Response_set-class}} object}
\usage{
\S4method{print}{Response_set}(x, ...)
}
\arguments{
\item{x}{An \code{\link{Response_set-class}} object that will be showed.}

\item{...}{Additional arguments. For example, an argument \code{n = 14},
will print 14 items to the console.}
}
\description{
Print a \code{\link{Response_set-class}} object
}
\author{
Emre Gonulates
}
\keyword{internal}
