% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilog.R
\name{print.bilog_output}
\alias{print.bilog_output}
\title{Prints \code{bilog_output} objects}
\usage{
\method{print}{bilog_output}(x, ...)
}
\arguments{
\item{x}{A \code{bilog_output} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints \code{bilog_output} objects
}
\author{
Emre Gonulates
}
\keyword{internal}
