% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_sim_helper_functions.R
\name{print.cat_design}
\alias{print.cat_design}
\title{Prints cat_design objects.}
\usage{
\method{print}{cat_design}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{A \code{cat_design} object.}

\item{...}{further arguments passed to or from other methods.}

\item{verbose}{If \code{TRUE}, a list object will be returned for each
step.}
}
\description{
Prints cat_design objects.
}
\examples{
ip <- generate_ip(n = 5)
cd <- create_cat_design(ip = ip, next_item_rule = 'random',
                        termination_rule = 'min_item',
                        termination_par = list('min_item' = 5))
cd
}
\author{
Emre Gonulates
}
\keyword{internal}
