% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-class-methods.R
\name{print.Item}
\alias{print.Item}
\title{Print an \code{\link{Item-class}} object}
\usage{
print.Item(x, ...)
}
\arguments{
\item{x}{An \code{\link{Item-class}} object to be printed.}

\item{...}{Passed parameters.}
}
\description{
Print an \code{\link{Item-class}} object
}
\author{
Emre Gonulates
}
\keyword{internal}
