\name{pwts1}
\alias{pwts1}
\title{pwts1}
\description{ pwts1 is a quadrature scheme for piecewise integration integrating from 0 to 1 and want total integral and each piece-this routine gives concatenation of evaluation points and weights }
\arguments{
   \item{zPoints}{ zPoints unevenly spaced over (0,1). }
   \item{nqpts}{ nqpts }
}
\value{ Function returns a... YW ... }
\author{Richard Spady}
\examples{}   
\keyword{}