% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\title{Rasch model}
\description{
Rasch model
}
\section{Slots}{

\describe{
\item{\code{b}}{Item difficulty parameter}

\item{\code{se_b}}{Standard error of item difficulty parameter}
}}

\author{
Emre Gonulates
}
