% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-class.R
\docType{class}
\name{Response-class}
\alias{Response-class}
\title{An S4 class representing responses of a single examinee}
\description{
An S4 class representing responses of a single examinee
}
\section{Slots}{

\describe{
\item{\code{examinee_id}}{Examinee/Subject/Student ID. A string or an integer to
identify an examinee.}

\item{\code{item_id}}{A character vector holding the item IDs.}

\item{\code{testlet_id}}{A character vector holding the testlet IDs that given item
belongs. It can be \code{NULL} if none of the items belongs to any testlet.
Items that do not belong to any testlet should be represented by \code{NA}.}

\item{\code{score}}{A numeric vector holding the scores given to items.}

\item{\code{raw_response}}{A vector of strings holding the raw responses to items.}

\item{\code{order}}{An integer vector representing the administration order of an
item.}

\item{\code{response_time}}{A numeric vector representing the response times. By
default, numbers are assumed to represent seconds.}

\item{\code{misc}}{A list that will hold miscellaneous information about the
responses. For example,
\code{misc = list(item_role = c("O", "O", "O", "F"))} will hold whether
administered item is a field test or an operational test item.}
}}

\author{
Emre Gonulates
}
