% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rmsd}
\alias{rmsd}
\title{Calculate Root Mean Square Deviation (RMSD)
 (or Root Mean Square Error (RMSE))}
\usage{
rmsd(est, true)
}
\arguments{
\item{est}{A numeric vector of estimated values}

\item{true}{A numeric vector of true values}
}
\value{
A number representing RMSD.
}
\description{
The formula for RMSD is:

\deqn{RMSD = \sqrt{\frac{\sum_{i = 1}^n(\hat \theta - \theta)^2}{n}}}
}
\examples{
true <- rnorm(10)
est <- true + runif(10)

}
\author{
Emre Gonulates
}
\keyword{internal}
