% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsss.R
\name{rsss}
\alias{rsss}
\title{Convert raw score to scale score and vice versa}
\usage{
rsss(ip, raw_score = NULL, scale_score = NULL, theta_range = c(-5, 5))
}
\arguments{
\item{ip}{An \code{\link{Itempool-class}} object.}

\item{raw_score}{A value (or vector of values) representing raw score(s).}

\item{scale_score}{A value (or vector of values) representing scale score(s).}

\item{theta_range}{The limits of the scale score. The default is
\code{c(-5, 5)}.}
}
\value{
A vector of raw or scale scores.
}
\description{
Convert raw score to scale score and vice versa
}
\author{
Emre Gonulates
}
