% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_analysis.R
\name{score_raw_resp}
\alias{score_raw_resp}
\title{Score Raw Responses}
\usage{
score_raw_resp(raw_resp, key)
}
\arguments{
\item{raw_resp}{A \code{matrix} or \code{data.frame} containing the item
responses.}

\item{key}{The answer key for the responses}
}
\value{
Scored response matrix
}
\description{
Score Raw Responses
}
\author{
Emre Gonulates
}
\keyword{internal}
