% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_sim_helper_functions.R
\name{show.cat_output}
\alias{show.cat_output}
\title{This method shows an "cat_output" class object}
\usage{
show.cat_output(object)
}
\arguments{
\item{object}{An 'cat_output' class object that will be showed.}
}
\description{
This method shows an "cat_output" class object
}
\author{
Emre Gonulates
}
\keyword{internal}
