% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-class-methods.R,
%   R/itempool-class-methods.R, R/response-class-methods.R,
%   R/response_set-class-methods.R, R/testlet-class-methods.R
\name{show,Item-method}
\alias{show,Item-method}
\alias{show,Itempool-method}
\alias{show,Response-method}
\alias{show,Response_set-method}
\alias{show,Testlet-method}
\title{Show an \code{\link{Item-class}} object}
\usage{
\S4method{show}{Item}(object)

\S4method{show}{Itempool}(object)

\S4method{show}{Response}(object)

\S4method{show}{Response_set}(object)

\S4method{show}{Testlet}(object)
}
\arguments{
\item{object}{A \code{\link{Testlet-class}} object that will be showed.}
}
\description{
Show an \code{\link{Item-class}} object

Show an \code{\link{Itempool-class}} object

Show an \code{\link{Response-class}} object

Show a \code{\link{Response_set-class}} object

Show a \code{\link{Testlet-class}} object
}
\author{
Emre Gonulates
}
\keyword{internal}
