% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_set-class-methods.R
\name{[[,Response_set,numeric,missing-method}
\alias{[[,Response_set,numeric,missing-method}
\title{Subset \code{Response_set} objects}
\usage{
\S4method{[[}{Response_set,numeric,missing}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{\link{Response_set-class}} object from which to extract
Response(s) or in which to replace element(s).}

\item{i}{indices specifying elements to extract or replace.}

\item{j}{This will not be used in \code{\link{Response_set-class}} objects.}

\item{...}{Additional parameters to be passed to the function.}
}
\value{
A \code{\link{Response-class}} object.
}
\description{
Subset \code{Response_set} objects
}
\examples{
resp_set <- sim_resp(ip = generate_ip(n = 12), theta = rnorm(10),
                     output = "response_set")
resp_set[[2]]
}
\author{
Emre Gonulates
}
\keyword{internal}
