% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testlet-class-methods.R
\name{[[,Testlet,numeric,missing-method}
\alias{[[,Testlet,numeric,missing-method}
\title{Access the items of a \code{\link{Testlet-class}} object.}
\usage{
\S4method{[[}{Testlet,numeric,missing}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{Testlet}  object from which to extract element(s) or in
which to replace element(s).}

\item{i}{indices specifying elements to extract or replace.}

\item{j}{This will not be used in \code{\link{Testlet-class}} objects.}

\item{...}{Additional parameters to be passed to the function.}
}
\value{
An  object with elements from 'Item' class.
}
\description{
Access the items of a \code{\link{Testlet-class}} object.
}
\author{
Emre Gonulates
}
\keyword{internal}
