% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_set-class-methods.R
\name{[[<-,Response_set,numeric,missing-method}
\alias{[[<-,Response_set,numeric,missing-method}
\title{Set the elements of an \code{Response_set} objects.}
\usage{
\S4method{[[}{Response_set,numeric,missing}(x, i, j) <- value
}
\arguments{
\item{x}{\code{\link{Response-class}} object.}

\item{i}{indices specifying elements to extract or replace.}

\item{j}{This will not be used in \code{\link{Response_set-class}} objects.}

\item{value}{An \code{\link{Response-class}} object.}

\item{...}{Additional parameters to be passed to the function.}
}
\value{
An updated \code{\link{Response_set-class}} object.
}
\description{
Set the elements of an \code{Response_set} objects.
}
\examples{

resp_set <- sim_resp(ip = generate_ip(n = 12), theta = rnorm(10),
                     output = "response_set")
r1 <- response(score = c(0, 1), examinee_id = "Stu-22",
               item_id = c("Item_1", "Item_2"))
resp_set[[2]] <- r1
}
\author{
Emre Gonulates
}
\keyword{internal}
