% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testlet-class-methods.R
\name{[[<-,Testlet,numeric,missing-method}
\alias{[[<-,Testlet,numeric,missing-method}
\title{This function sets the elements of a Testlet objects.}
\usage{
\S4method{[[}{Testlet,numeric,missing}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{Testlet} object from which to extract element(s) or
in which to replace element(s).}

\item{i}{indices specifying elements to extract or replace.}

\item{j}{This will not be used in \code{\link{Testlet-class}} objects.}

\item{value}{An \code{Item} object.}

\item{...}{Additional parameters to be passed to the function.}
}
\value{
An \code{teslet} object with elements from 'Item' class.
}
\description{
This function sets the elements of a Testlet objects.
}
\author{
Emre Gonulates
}
\keyword{internal}
