% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testlet-class-methods.R
\name{[,Testlet,ANY,missing-method}
\alias{[,Testlet,ANY,missing-method}
\title{Subset \code{\link{Testlet-class}} object}
\usage{
\S4method{[}{Testlet,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{Testlet} object from which to extract element(s) or in
which to replace element(s).}

\item{i}{indices specifying elements to extract or replace.}

\item{j}{This will not be used in \code{\link{Testlet-class}} objects.}

\item{...}{Parameters to be passed to the function.}

\item{drop}{(From R manual:) For matrices and arrays. If TRUE the result is
coerced to the lowest possible dimension (see the examples). This only works
for extracting elements, not for the replacement. See drop for further
details.}
}
\value{
An list object with elements from 'Item' class.
}
\description{
Subset \code{\link{Testlet-class}} object
}
\author{
Emre Gonulates
}
\keyword{internal}
