% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var,Item-method}
\alias{var,Item-method}
\alias{var,Rasch-method}
\alias{var,1PL-method}
\alias{var,2PL-method}
\alias{var,3PL-method}
\alias{var,4PL-method}
\alias{var,GRM-method}
\alias{var,PCM-method}
\alias{var,GPCM-method}
\alias{var,GPCM2-method}
\title{Calculate the variance of an Item}
\usage{
\S4method{var}{Item}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{Rasch}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{`1PL`}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{`2PL`}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{`3PL`}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{`4PL`}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{GRM}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{PCM}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{GPCM}(x, y = NULL, na.rm = FALSE, use)

\S4method{var}{GPCM2}(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{An \code{\link{Item-class}} or an \code{\link{Itempool-class}}
object containing the item parameters.}

\item{y}{A numeric vector containing the ability parameters (i.e. theta).}

\item{na.rm}{Ignored for \code{var(Item, ...)}}

\item{use}{Ignored for \code{var(Item, ...)}}
}
\value{
Item variances at given theta will be returned.
}
\description{
\code{var} Returns the variance of  an item or multiple items
  with given parameters for a given ability or abilities, i.e. \eqn{\theta}.
}
\author{
Emre Gonulates
}
