% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var,Itempool-method}
\alias{var,Itempool-method}
\title{Calculate the variances of items in an Itempool}
\usage{
\S4method{var}{Itempool}(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{An \code{\link{Itempool-class}} object containing the item
parameters.}

\item{y}{A numeric vector containing the ability parameters (i.e. theta).}

\item{na.rm}{Ignored for \code{var(Itempool, ...)}}

\item{use}{Ignored for \code{var(Itempool, ...)}}
}
\value{
Item variances at given theta will be returned.
}
\description{
\code{var} Returns the variance of each item of an
  \code{\link{Itempool-class}} object for a given ability or abilities,
  i.e. \eqn{\theta}.
}
\author{
Emre Gonulates
}
