% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var,Testlet-method}
\alias{var,Testlet-method}
\title{Calculate the variances of items in a Testlet}
\usage{
\S4method{var}{Testlet}(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{An \code{\link{Testlet-class}} object containing the item
parameters of the testlet.}

\item{y}{A numeric vector containing the ability parameters (i.e. theta).}

\item{na.rm}{Ignored for \code{var(Testlet, ...)}}

\item{use}{Ignored for \code{var(Testlet, ...)}}
}
\value{
Item variances at given theta will be returned.
}
\description{
Calculate the variances of items in a Testlet
}
\author{
Emre Gonulates
}
