% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class-methods.R
\name{$<-,Itempool-method}
\alias{$<-,Itempool-method}
\title{Set values to parameters or components of 'Item' class.}
\usage{
\S4method{$}{Itempool}(x, name) <- value
}
\arguments{
\item{x}{\code{\link{Itempool-class}} object.}

\item{name}{Name of the parameter or component. Currently only \code{misc},
\code{id}, \code{content}, \code{item_list} are available.}

\item{value}{The new value that will be assigned.
\itemize{
  \item For \code{id}, the value should be a list of strings that
        has the same length as the length of the
        \code{\link{Itempool-class}} object. There should not be
        any duplicated id's.
  \item For \code{content}, the value should be either \code{NULL}
        or a list of strings that has the same length as the
        length of the \code{\link{Itempool-class}} object.
  \item For \code{item_list}, the value should be a list of
        \code{\link{Item-class}} or \code{\link{Testlet-class}}
        objects.
  \item For \code{misc}, the value should be a list.
}}
}
\value{
This operation will return an \code{\link{Itempool-class}} object.
}
\description{
Set values to parameters or components of 'Item' class.
}
\author{
Emre Gonulates
}
