% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class-methods.R
\name{$<-,Testlet-method}
\alias{$<-,Testlet-method}
\title{Set values to parameters or components of 'Item' class.}
\usage{
\S4method{$}{Testlet}(x, name) <- value
}
\arguments{
\item{x}{A \code{\link{Testlet-class}} object.}

\item{name}{Name of the parameter or component.}

\item{value}{The new value that will be assigned.}
}
\value{
This operation will not return anything.
}
\description{
Set values to parameters or components of 'Item' class.
}
\author{
Emre Gonulates
}
