% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testlet-class-methods.R
\name{$,Testlet-method}
\alias{$,Testlet-method}
\title{Access slots of a \code{\link{Testlet-class}} object}
\usage{
\S4method{$}{Testlet}(x, name)
}
\arguments{
\item{x}{A \code{\link{Testlet-class}} object from which to extract
element(s) or in which to replace element(s).}

\item{name}{Name of the parameter.
Available values:
\describe{
  \item{\strong{\code{'id'}}}{Get the \code{id} of the testlet}
  \item{\strong{\code{'content'}}}{Get the \code{content} of
    the testlet.}
  \item{\strong{\code{'model'}}}{Get the \code{model} of the
    testlet.}
  \item{\strong{\code{'item_models'}}}{Get the \code{model} of the
    items within the testlet.}
  \item{\strong{\code{'parameters'}}}{Get the \code{parameters} of
    the testlet.}
  \item{\strong{\code{'se_parameters'}}}{Get the
    \code{se_parameters} of the testlets.}
  \item{\strong{\code{'item_list'}}}{Get the list of
    \code{\link{Item-class}} objects of the testlet. Returns a
    \code{list} object.}
  \item{\strong{\code{'max_score'}}}{Returns the maximum score
    obtainable by all of the items within the testlet.}
}}
}
\value{
This operation will return the desired slot.
}
\description{
Access slots of a \code{\link{Testlet-class}} object
}
\examples{
t1 <- testlet(generate_ip(n = 3), id = "my-testlet", content = "Algebra")
t1$model
t1$id
t1$item_list
t1$content
t1$item_models
}
