% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt.R
\docType{package}
\name{irt}
\alias{irt}
\title{A Collection of Item Response Theory (IRT) and Computerized Adaptive Testing
(CAT) Functions}
\description{
A collection of Item Response Theory (IRT) and Computerized
  Adaptive Testing (CAT) functions that are used in psychometrics.
}
\author{
Emre Gonulates \email{egonulates@gmail.com}
}
\keyword{package}
