% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Itempool-class-methods.R,
%   R/Testlet-class-methods.R
\name{length,Itempool-method}
\alias{length,Itempool-method}
\alias{length,Testlet-method}
\title{Find the length of an \code{\link{Itempool-class}} object}
\usage{
\S4method{length}{Itempool}(x)

\S4method{length}{Testlet}(x)
}
\arguments{
\item{x}{an \code{\link{Itempool-class}} object}
}
\description{
Find the length of an \code{\link{Itempool-class}} object

Find the length of a \code{\link{Testlet-class}} object
}
\author{
Emre Gonulates
}
