% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_fit.R
\name{person_fit}
\alias{person_fit}
\alias{person_fit,Item-method}
\alias{person_fit,Itempool-method}
\alias{person_fit,Testlet-method}
\title{Calculate person-fit indices}
\usage{
person_fit(ip, resp, theta, type = "lz")

\S4method{person_fit}{Item}(ip, resp, theta, type = "lz")

\S4method{person_fit}{Itempool}(ip, resp, theta, type = "lz")

\S4method{person_fit}{Testlet}(ip, resp, theta, type = "lz")
}
\arguments{
\item{ip}{An \code{\link{Item-class}}, \code{\link{Itempool-class}} or a
\code{\link{Testlet-class}} object.}

\item{resp}{A vector of item responses.}

\item{theta}{An vector containing ability parameters.}

\item{type}{The type of the person-fit index.}
}
\value{
A vector of person-fit index values.
}
\description{
\code{person_fit} calculates the fit of a person to a given psychometric
model.
}
\author{
Emre Gonulates
}
