% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_sim_helper_functions.R
\name{print.cat_output}
\alias{print.cat_output}
\title{This method prints an "cat_output" class object}
\usage{
\method{print}{cat_output}(x, ..., silent = FALSE)
}
\arguments{
\item{...}{Additional arguments}

\item{silent}{If TRUE, no output will be printed on the console, only
a data frame will be returned.}

\item{object}{An 'cat_output' class object that will be printed.}
}
\description{
This method prints an "cat_output" class object
}
\author{
Emre Gonulates
}
\keyword{internal}
