% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-class-methods.R,
%   R/Itempool-class-methods.R, R/Testlet-class-methods.R
\name{show,Item-method}
\alias{show,Item-method}
\alias{show,Itempool-method}
\alias{show,Testlet-method}
\title{Show an \code{\link{Item-class}} object}
\usage{
\S4method{show}{Item}(object)

\S4method{show}{Itempool}(object)

\S4method{show}{Testlet}(object)
}
\arguments{
\item{object}{A \code{\link{Testlet-class}} object that will be showed.}
}
\description{
Show an \code{\link{Item-class}} object

Show an \code{\link{Itempool-class}} object

Show a \code{\link{Testlet-class}} object
}
\author{
Emre Gonulates

Emre Gonulates

Emre Gonulates
}
\keyword{internal}
